move cidr_range_denylist from globals to client service
